+++ src/main/java/com/iexec/worker/feign/CustomCoreFeignClient.java
@@ -18,11 +18,11 @@ package com.iexec.worker.feign;
 
 import com.iexec.common.config.WorkerModel;
 import com.iexec.common.replicate.ReplicateStatusUpdate;
-import com.iexec.common.replicate.ReplicateTaskSummary;
 import com.iexec.core.api.SchedulerClient;
 import com.iexec.core.config.PublicConfiguration;
 import com.iexec.core.notification.TaskNotification;
 import com.iexec.core.notification.TaskNotificationType;
+import com.iexec.core.replicate.ReplicateTaskSummary;
 import feign.FeignException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;